Make SEO improvements: - Make pages server components - Move page content to their own client components - Add metadata descriptions and titles for each page
