+++ src/app/contact/page.tsx
@@ -1,5 +1,13 @@
+import { Metadata } from 'next';
+
 import ContactForm from '@/components/ContactForm/ContactForm';
 import Link from 'next/link';
+
+export const metadata: Metadata = {
+  title: 'Tag Operations | Contact',
+  description:
+    "Get in touch with Tag Ops to explore tailored solutions for your business. Whether large or small, we're here to support your operational goals. Fill out the form or email us at tag@tagoperations.com to start the conversation.",
+};
 
 export default function ContactPage() {
   return (
@@ -7,7 +15,7 @@ export default function ContactPage() {
       <section className="w-full flex flex-col items-center mt-8 p-6 pb-2 md:w-1/2 lg:w-1/3 md:items-start">
         <h2 className="text-4xl text-primary md:text-5xl">Contact</h2>
 
-        <div className='flex flex-col'>
+        <div className="flex flex-col">
           <p className="mt-2 md:text-lg md:mt-4">
             Fill out this form to start the conversation on how we can best
             support your business. Whether your business is large or small,
